+++ backend/handlers/graphqlHandler.js
@@ -8,7 +8,7 @@ const pgClientFromContext = require("postgraphql/build/postgres/inventory/pgClie
 const setupRequestPgClientTransaction = require("postgraphql/build/postgraphql/http/setupRequestPgClientTransaction");
 const PgCat = JSON.parse(fs.readFileSync('pgCatalog/pgCatalog.json', 'utf8'));
 
-module.exports.graphql = (event, context, cb) => {
+module.exports.handler = (event, context, cb) => {
   // Setup connection to PostgresDB
   const pgClient = new Client(process.env.PGCON);
   pgClient.connect();
@@ -66,24 +66,3 @@ module.exports.graphql = (event, context, cb) => {
         cb(null, {"errors": [err]});
       });
     }
-
-module.exports.xml = (event, context, cb) => {
-  const pgClient = new Client(process.env.PGCON);
-  pgClient.connect();
-
-  pgClient.query('select floods.legacy_xml()')
-    .then(pgres => 
-    {
-      const response = {
-          statusCode: 200,
-          headers: {
-            'Content-Type': 'text/xml',
-          },
-          body: pgres.rows[0].legacy_xml,
-      };
-
-      cb(null, response);
-    })
-    .catch(err => cb(null, {"errors": [err]}))
-    .then(() => pgClient.end());    
-}
